Enabling continuous die rolls & average prints
Rewrote the code to allow the user to roll again, which required a lot of new code and pretty some tweaking of what I already wrote.
First change is running the Testing object first, before the Game object. This just makes things more straightforward.
Added a game loop to Program, which loops until the user inputs "N"/"n" when asked if they want to roll again. This loop has it's own validation loop that loops until it gets a valid input.
Added new properties to Game - _games and _average, which saves the number of games and sum (mean) average respectively. These will be printed when the game loop is ended, to take advantage of continuous dice rolls.
This average calculation is implemented by adding a new if/else if/else statement to the end of RollDie(), which updates the aforementioned new values.
Added a metric ton of new comments to the code for these changes.
What I still need to do is update Testing to at least somewhat acknowledge these changes. That should wrap up my work on this assignment code-wise.