Updated Testing for new code + minor tweaks
Added a new property to Testing to store the number of games ran.
Tested the default values of the "Games" and "Average" properties in the same way as the others.
Amended the Debug.Assert()s to use AND statements instead of OR statements, because that was a mistake, oops!!!
Tested the "Games" and "Average" properties after one game, and then also after ten games.
Doing this required removing the code from the Game RollDie() method that printed values, because printing all that test data would be weird. That's now separate!! Whey
I also did some minor things:
-Averages now round to two decimal points
-I made it print the exception on the try/catch statement
-I made it use that one *= operator at one point