Implemented (mostly) Three Or More
Coded the ruleset for Three Or More, which requires a great deal of variables and at least two for loops inside each other.
Also did more coding for the main menu, specifically calling the two respective games.
Commented out most of what is left of Game.cs. May end up deleting it now that it's obsolete.